feat: ファイル分けた、setup.pyを設定した
