+++ what_recent/__init__.py
@@ -1,0 +1,2 @@
+from .event import EventFactory
+from .summary import activities_summary
