+++ what_recent/event.py
@@ -1,9 +1,3 @@
-import json
-import openai
-import requests
-import sys
-import os
-
 BOT_USER_NAMES=['renovate[bot]', 'dependabot[bot]', 'github-actions[bot]']
 
 class EventFactory:
@@ -149,62 +143,4 @@ class ForkEvent(Event):
 
     def available(self):
         return True
-
-
-def read_json(filename):
-    with open(filename) as f:
-        return json.load(f)
-
-def activities_summary(activities):
-    model = os.getenv('OPENAI_MODEL', 'gpt-3.5-turbo')
-    system_content = """
-これからGitHubでのリポジトリとどんなことをしたかのテキストを渡します
-これらをまとめて、最近の活動についてまとめてください
-どういう技術に興味がありそうなのか、取り組みの難易度もわかれば教えてほしいです
-同じようなことをたくさんしている場合はその旨も要約に入れてください
-"""
-
-    text = "\n".join(list(map(lambda a: f"リポジトリ:{a['repository']}\n{a['changes']}", activities)))
-
-    # print("input ==================================================")
-    # print(text)
-    # print("input ==================================================")
-
-    system_prompt = {"role": "system", "content": system_content}
-
-    res = openai.ChatCompletion.create(
-        model=model,
-        temperature=0.9,
-        messages=[
-            system_prompt,
-            {"role": "user", "content": text }
-        ])
-
-    role = res.choices[0]["message"]["role"]
-    content = res.choices[0]["message"]["content"]
-
-    print(f"{role}: {content}")
-
-    return content
-
-
-user = sys.argv[1]
-url = f"https://api.github.com/users/{user}/events?per_page=100&page=1"
-res = requests.get(url)
-data = res.json()
-factory = EventFactory()
-
-events = [factory.create_event(event) for event in data]
-activities = [e.transform() for e in events if e.available()]
-dummy = [e for e in events if e.__class__.__name__ == 'DummyEvent']
-bot_activities = [e for e in events if e.bot_event()]
-
-# print(set([e.event_name for e in dummy]))
-# print(json.dumps([e.data for e in dummy if e.event_name == 'ForkEvent'], indent=2))
-
-print(f"{len(activities)}件のサマリー対象Activityがありました")
-print(f"{len(dummy)}件の未対応Activityがありました")
-print(f"{len(bot_activities)}件のbotによるActivityがありました")
-
-activities_summary(activities)
